requirements.txt
pystray
wmi
pywin32
pillow